Add DebugLogger middleware
It is useful to have a middleware that can be placed at any (or all)
points in the middleware stack, and which can report on values passing
up or down the stack.

Due to the fact that an app returns an iterator, this logger can't
necessarily report on the values being returned at the point that they
are passed back up the middleware stack, as that would involve enumerating,
and hence potentially demolishing, the contents of the iterator. Instead,
values are reported on at the point of use by using a wrapping iterator
which does the logging instead.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>