Output UTF8 bytes, not Python strings
Everything that gets sent to (or read from) the outside world in py3 needs
to be treated as an encoded byte sequence, not as a python string. This
shouldn't be done universally, e.g. as a side-effect of some middleware,
as then it's hard to override. In fact, the conneg layer should be doing
this, as it's almost entirely responsible for what gets sent back to the
requester.

So, the trivially-stupid default conneg transformer should do the
trivially-stupid thing and convert everything to UTF-8.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>