Fix content negotiation
The original implementation of content negotiation tried to determine the
transformer too early in the process -- we could make a decision about the
requested content type, but couldn't turn that into a suitable transformer
class, because it hadn't been set by the downstream app code.

Moving the decision into the start_response call means that we've already
run the app code when we get to it, so we should have a "transformers"
dictionary to look in by that point.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>