Use ts_type throughout the application, to avoid confusion
The allowable parameter names in get_series_list are checked for matches,
and then passed directly to the database as a function parameter name.
Thus, using "type" in the public API and "ts_type" in the private DB API
is a bit of a mismatch. We can either hack it to match everywhere in the
fuse.api module, or refactor the name to ts_type everywhere. This patch
performs the latter operation.

Signed-off-by: Hugo Mills <h.r.mills@reading.ac.uk>