Add metadata structures to data retrieval
When retrieving multiple data series, we want to be able to see which
column refers to which data set. This patch modifies the output of
the get_data() functions to return a lump of metadata as well as the data
itself. The data is also now converted to a tabular form.

Signed-off-by: Hugo Mills <h.r.mills@reading.ac.uk>